+++ server.js
@@ -1,8 +1,10 @@
 const express = require("express")
-const cookieParser = require("cookie-parser")
-const path = require("path")
 const app = express()
 const port = 3000
+
+const path = require("path")
+const cookieParser = require("cookie-parser")
+const expressSession = require('express-session');
 
 require("dotenv").config()
 
@@ -11,10 +13,14 @@ const followApi = require("./router/follow.services.js")
 const profileProviedApi = require("./router/profile.provide.js")
 const profileServicesApi = require("./router/profile.services.js")
 const acconutApi = require('./router/account.provider')
+const kakaoRegisterApi = require('./router/kakao.account.js');
+const createSessionConfig = require('./config/session');
 
 // const tapeApi = require("./router/tape.services")
 
 // app.use ---------------------------------------------------
+const sessionConfig = createSessionConfig();
+app.use(expressSession(sessionConfig));
 app.use(express.urlencoded({extended:false}));
 app.use(express.json())
 
@@ -25,6 +31,7 @@ app.use("/follow", followApi)
 app.use("/profile", profileProviedApi)
 app.use("/profile", profileServicesApi)
 app.use('/account', acconutApi);
+app.use("/kakao", kakaoRegisterApi);
 // app.use("/tape", tapeApi)
 
 // app.get ---------------------------------------------------
